fix: move component, add better empty tate for member policy
